refactor: 테스트 코드를 프로덕션 코드 패키지 구조와 동일하게 변경
