+++ backend/src/main/java/mouda/backend/moim/presentation/controller/ChatController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RestController;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import mouda.backend.aop.logging.ExceptRequestLogging;
-import mouda.backend.common.response.RestResponse;
 import mouda.backend.common.config.argumentresolver.LoginDarakbangMember;
+import mouda.backend.common.response.RestResponse;
 import mouda.backend.darakbangmember.domain.DarakbangMember;
 import mouda.backend.moim.business.ChatService;
 import mouda.backend.moim.presentation.controller.swagger.ChatSwagger;
