+++ backend/src/main/java/mouda/backend/moim/presentation/controller/swagger/ChatSwagger.java
@@ -8,8 +8,8 @@ import org.springframework.web.bind.annotation.RequestParam;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import mouda.backend.common.response.RestResponse;
 import mouda.backend.common.config.argumentresolver.LoginDarakbangMember;
+import mouda.backend.common.response.RestResponse;
 import mouda.backend.darakbangmember.domain.DarakbangMember;
 import mouda.backend.moim.presentation.request.chat.ChatCreateRequest;
 import mouda.backend.moim.presentation.request.chat.DateTimeConfirmRequest;
