+++ backend/src/main/java/mouda/backend/moim/presentation/controller/swagger/ZzimSwagger.java
@@ -9,8 +9,8 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import jakarta.validation.Valid;
-import mouda.backend.common.response.RestResponse;
 import mouda.backend.common.config.argumentresolver.LoginDarakbangMember;
+import mouda.backend.common.response.RestResponse;
 import mouda.backend.darakbangmember.domain.DarakbangMember;
 import mouda.backend.moim.presentation.request.zzim.ZzimUpdateRequest;
 import mouda.backend.moim.presentation.response.zzim.ZzimCheckResponse;
