+++ backend/src/test/java/mouda/backend/common/fixture/MemberFixture.java
@@ -1,0 +1,24 @@
+package mouda.backend.common.fixture;
+
+import mouda.backend.member.domain.Member;
+
+public class MemberFixture {
+
+	public static Member getHogee() {
+		return Member.builder()
+			.nickname("hogee")
+			.build();
+	}
+
+	public static Member getAnna() {
+		return Member.builder()
+			.nickname("anna")
+			.build();
+	}
+
+	public static Member getTebah() {
+		return Member.builder()
+			.nickname("tebah")
+			.build();
+	}
+}
