+++ backend/src/test/resources/META-INF/services/org.junit.jupiter.api.extension.Extension
@@ -1,1 +1,1 @@
-mouda.backend.config.NoTransactionExtension
+mouda.backend.common.config.NoTransactionExtension
